+++ packages/Peptides/src/widgets/distribution.ts
@@ -167,7 +167,9 @@ export function getDistributionWidget(table: DG.DataFrame, model: PeptidesModel)
         const mask = DG.BitSet.create(rowCount,
           bitArray.allFalse || bitArray.allTrue ? (_): boolean => true : (i): boolean => bitArray.getBit(i));
         const aggregatedColMap = model.getAggregatedColumnValues({filterDf: true, mask});
-        const stats = bitArray.allFalse || bitArray.allTrue ? {count: rowCount, pValue: null, meanDifference: 0, ratio: 1, mask: bitArray} :
+        const stats = bitArray.allFalse || bitArray.allTrue ?
+          {count: rowCount, pValue: null, meanDifference: 0, ratio: 1, mask: bitArray,
+            mean: activityCol.stats.avg} :
           getStats(activityColData, bitArray);
         const tableMap = getStatsTableMap(stats);
         const resultMap: {[key: string]: any} = {...tableMap, ...aggregatedColMap};